requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
openai
google-generativeai
python-dateutil
pendulum
httpx
